HTML and Markdown fix
There appears to be a bug in Qt that causes HTML and Markdown output to only produce level 1 headers if text is appended to a QTextEdit.

The following is a work-around for that bug, by setting the entire output of the QTextEdit by concatenating the new text onto the current text, rather than just doing a simple append operation. It also sets the location of the cursor to the end of the edit.